Remove length check
1. No attempt to be backwards compatible at the moment
2. This also implies that "long" can be used instead of "Long" (since null checking isnt needed)

+++ test/integration/src/test/java/io/pravega/test/integration/AppendTest.java
@@ -128,7 +128,7 @@ public class AppendTest {
                                                       new Append(segment, uuid, data.readableBytes(), data, null));
         assertEquals(uuid, ack.getWriterId());
         assertEquals(data.readableBytes(), ack.getEventNumber());
-        assertEquals(Long.MIN_VALUE, (long) ack.getPreviousEventNumber());
+        assertEquals(Long.MIN_VALUE, ack.getPreviousEventNumber());
     }
 
     @Test(timeout = 10000)
@@ -153,13 +153,13 @@ public class AppendTest {
                 new Append(segment, uuid, 1, data, null));
         assertEquals(uuid, ack.getWriterId());
         assertEquals(1, ack.getEventNumber());
-        assertEquals(Long.MIN_VALUE, (long) ack.getPreviousEventNumber());
+        assertEquals(Long.MIN_VALUE, ack.getPreviousEventNumber());
 
         DataAppended ack2 = (DataAppended) sendRequest(channel,
                 new Append(segment, uuid, 2, data, null));
         assertEquals(uuid, ack2.getWriterId());
         assertEquals(2, ack2.getEventNumber());
-        assertEquals(1, (long) ack2.getPreviousEventNumber());
+        assertEquals(1, ack2.getPreviousEventNumber());
     }
